failed attempt at rendering the ZE widget in Layout
