+++ src/pages/_app.tsx
@@ -7,10 +7,7 @@ import { CacheProvider } from "@emotion/react";
 import * as Toast from "@radix-ui/react-toast";
 import * as Tooltip from "@radix-ui/react-tooltip";
 import type { AppProps } from "next/app";
-import { useRouter } from "next/router";
 import { ThemeProvider } from "next-themes";
-import { useEffect } from "react";
-import Zendesk, { ZendeskAPI } from "react-zendesk";
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import tw from "twin.macro";
 
@@ -20,14 +17,12 @@ import {
   LayoutApp,
   UserProvider,
 } from "src/components";
-import config from "src/config";
 
 import GlobalStyles from "../styles/GlobalStyles";
 import { useApollo } from "../utils/apolloClient";
 
 const NextApp = ({ Component, pageProps }: AppProps) => {
   const client = useApollo(pageProps);
-  const router = useRouter();
 
   // Datadog RUM initialization
   datadogRum.init({
@@ -44,21 +39,6 @@ const NextApp = ({ Component, pageProps }: AppProps) => {
   });
   datadogRum.startSessionReplayRecording();
 
-  useEffect(() => {
-    const hideZendesk = config.routesToHideZendeskWidget.some((path) =>
-      router.pathname.startsWith(path),
-    );
-    console.log("hideZendesk", hideZendesk);
-
-    if (hideZendesk) {
-      ZendeskAPI("webWidget", "hide");
-    }
-
-    return () => {
-      ZendeskAPI("webWidget", "show");
-    };
-  }, []);
-
   return (
     <ApolloProvider client={client}>
       <CacheProvider value={cache}>
@@ -71,11 +51,6 @@ const NextApp = ({ Component, pageProps }: AppProps) => {
                   <AppHook>
                     <LayoutApp renderNavMobile>
                       <Component {...pageProps} />
-                      <Zendesk
-                        defer
-                        zendeskKey={config.ZENDESK_WIDGET_KEY}
-                        onLoaded={() => console.log("ZE is loaded")}
-                      />
                     </LayoutApp>
                   </AppHook>
                 </AuthenticatedLayout>
